requirements.txt
django==3.0.5
model_bakery==1.1.0
Pillow==7.1.2
djangorestframework==3.11.0
django-mptt==0.11.0
